Program is functioning correctly
This program creates a custom linked list that has insert, delete, and iterate functions.  Program is working as it should.  It takes input from a text file, input.txt and inserts to into the custom linked list. It demonstrates the delete functionality by deleting the fist instance of 2.  The input.txt file contains the numbers 1 through 10 and the word hello to trigger an error. There are screenshots of the input and output of the program.  Also included is the pseudocode that was used to write the program.